change to double

+++ Calculator/Calculator.cs
@@ -2,15 +2,15 @@ namespace UtilityKnife.Calculator
 {
     public class Calculator
     {
-        public static int Add (int x, int y)
+        public static double Add (double x, double y)
         {
             return x + y;
         }
-        public static int Subtract(int x, int y)
+        public static double Subtract(double x, double y)
         {
             return x - y;
         }
-        public static int Multiply(int x, int y)
+        public static double Multiply(double x, double y)
         {
             return x * y;
         }
